Move old callbacks to legacy module; add new piecewise linear learning rate callback along with custom progress bar
